Rename myCoursesPage.ts to my-bluedot.ts
