+++ apps/website/src/server/routers/my-bluedot.ts
@@ -18,14 +18,12 @@ import {
   unitTable,
 } from '@bluedot/db';
 import type { CourseListRowProps } from '../../components/my-courses/CourseListRow';
-import { TRPCError } from '@trpc/server';
-import { z } from 'zod';
 import db from '../../lib/api/db';
 import { protectedProcedure, router } from '../trpc';
 import { getAvailableGroupsAndDiscussions } from './group-switching';
 
-export const myCoursesPageRouter = router({
-  getOverview: protectedProcedure.query(async ({ ctx }) => {
+export const myBluedotRouter = router({
+  myCoursesPage: protectedProcedure.query(async ({ ctx }) => {
     const { email } = ctx.auth;
 
     // Step 1: Fetch data
@@ -245,52 +243,4 @@ export const myCoursesPageRouter = router({
 
     return { courses: perCourse, nextDiscussion };
   }),
-
-  // Returns the people in the caller's group for a given meetPerson record. The caller must own
-  // the meetPerson (verified via email match). Facilitators are returned first, then participants
-  // alphabetically by name. Caller themselves are excluded from the list.
-  getGroupParticipants: protectedProcedure
-    .input(z.object({ meetPersonId: z.string().min(1) }))
-    .query(async ({ input, ctx }) => {
-      const meetPersonRows = await db.pg.select()
-        .from(meetPersonTable.pg)
-        .where(eq(meetPersonTable.pg.id, input.meetPersonId));
-      const callerMeetPerson = meetPersonRows[0];
-      if (callerMeetPerson?.email !== ctx.auth.email) {
-        throw new TRPCError({ code: 'NOT_FOUND', message: 'meetPerson not found' });
-      }
-
-      const groupId = callerMeetPerson.groupsAsParticipant?.[0];
-      if (!groupId) return { facilitators: [], participants: [] };
-
-      const groupRows = await db.pg.select().from(groupTable.pg).where(eq(groupTable.pg.id, groupId));
-      const group = groupRows[0];
-      if (!group) return { facilitators: [], participants: [] };
-
-      const facilitatorIdsArr = group.facilitator ?? [];
-      const participantIdsArr = (group.participants ?? []).filter((id) => id !== callerMeetPerson.id);
-      const allIds = [...facilitatorIdsArr, ...participantIdsArr];
-      if (allIds.length === 0) return { facilitators: [], participants: [] };
-
-      const people = await db.pg
-        .select({
-          id: meetPersonTable.pg.id,
-          name: meetPersonTable.pg.name,
-        })
-        .from(meetPersonTable.pg)
-        .where(inArray(meetPersonTable.pg.id, allIds));
-      const peopleById = new Map(people.map((p) => [p.id, { id: p.id, name: p.name ?? '' }]));
-
-      const facilitators = facilitatorIdsArr
-        .map((id) => peopleById.get(id))
-        .filter((p): p is { id: string; name: string } => !!p)
-        .sort((a, b) => a.name.localeCompare(b.name));
-
-      const participants = participantIdsArr
-        .map((id) => peopleById.get(id))
-        .filter((p): p is { id: string; name: string } => !!p)
-        .sort((a, b) => a.name.localeCompare(b.name));
-
-      return { facilitators, participants };
-    }),
 });
